Test 2 - testPasswordEncoding
