+++ src/test/java/sample/services/UserDataServiceTest.java
@@ -2,7 +2,9 @@ package sample.services;
 import org.json.simple.JSONArray;
 import org.junit.Test;
 
-import static org.junit.Assert.assertNotNull;
+import java.security.NoSuchAlgorithmException;
+
+import static org.junit.Assert.*;
 
 public class UserDataServiceTest {
     @Test
@@ -10,5 +12,11 @@ public class UserDataServiceTest {
         JSONArray jrr = UserDataService.OpenFile("UserData.json");
         assertNotNull(jrr);
     }
+    @Test
+    public void testPasswordEncoding() throws NoSuchAlgorithmException {
+        assertEquals("afa939adf52ddcbd204c814afcdd754a", UserDataService.hash("daa"));
+
+    }
+
 
 }
